// apps/pokeverse/src/helpers/types.ts
export type DifficultyId =
    | 'beginner' | 'intermediate' | 'advanced' | 'master';
// ...
export interface BattleMove {
    id: number;
    name: string;
    type: string;
    power: number;
    accuracy: number;
    category: 'Physical' | 'Special';
}

export interface BattlePokemon {
    id: number;
    name: string;
    types: string[];
    level: number;
    currentHP: number;
    maxHP: number;
    selectedMoves: BattleMove[];
    backSprite: string;
    frontSprite: string;
    calculatedStats: {
        hp: number;
        attack: number;
        defense: number;
        specialAttack: number;
        specialDefense: number;
        speed: number;
    };
}